requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.3
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
zipp==3.23.0
# Optional: desktop wrapper for the app
pywebview==3.8
